:construction: prog lab 5 done everything except exec_scr & save/load commands

+++ Programming/2 sem/Lab5/lab5/src/Modules/CommandHandler.java
@@ -3,11 +3,8 @@ package Modules;
 
 import CollectionObject.VehicleType;
 
-import java.io.File;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.LinkedList;
-import java.util.Scanner;
+
 
 public class CommandHandler {
     private CollectionService collectionService;
@@ -22,7 +19,7 @@ public class CommandHandler {
 
     public void help(String arguments){
         if (!arguments.isBlank()){
-            System.out.println("гавно дибил");// illegal args exception
+            System.out.println("Неверные аргументы команды");// illegal args exception
         } else {
             System.out.println(
                 """
@@ -50,10 +47,9 @@ public class CommandHandler {
         }
     }
 
-    // TODO: 19.02.2024 добавить исключения на аргументы
     public void info(String arguments){
         if (!arguments.isBlank()){
-            System.out.println("гавно дибил"); // illegal args exception
+            System.out.println("Неверные аргументы команды"); // illegal args exception
         } else {
             collectionService.info();
         }
@@ -61,7 +57,7 @@ public class CommandHandler {
 
     public void show(String arguments){
         if (!arguments.isBlank()){
-            System.out.println("гавно дибил"); // illegal args exception
+            System.out.println("Неверные аргументы команды"); // illegal args exception
         } else {
             collectionService.show();
         }
@@ -69,7 +65,7 @@ public class CommandHandler {
 
     public void add(String arguments){
         if (!arguments.isBlank()){
-            System.out.println("гавно дибил"); // illegal args exception
+            System.out.println("Неверные аргументы команды"); // illegal args exception
         } else {
             collectionService.addElement();
         }
@@ -77,28 +73,42 @@ public class CommandHandler {
 
     public void update(String arguments){  //args required
         if (arguments.isBlank()){
-            System.out.println("гавно дибил"); // illegal args exception
+            System.out.println("Неверные аргументы команды"); // illegal args exception
         } else {
-            String [] parsedArgs = arguments.split(" ");
-            long current_id = Long.parseLong(parsedArgs[0]);
-            String new_name = parsedArgs[1];
+            try {
+                long current_id = Long.parseLong(arguments);
+                if (current_id > 0){
+                    collectionService.update(current_id);
+                } else {
+                    System.out.println("id не может быть отрицательным");
+                }
 
-            collectionService.update(new_name, current_id);
+            } catch (NumberFormatException e){
+                System.out.println("Неверный формат аргументов");
+            }
         }
     }
 
     public void removeById(String arguments){ //args required
         if (arguments.isBlank()){
-            System.out.println("гавно дибил"); // illegal args exception
+            System.out.println("Неверные аргументы команды"); // illegal args exception
         } else {
-            long id = Long.parseLong(arguments);
-            collectionService.removeById(id);
+            try {
+                long id = Long.parseLong(arguments);
+                if (id > 0){
+                    collectionService.removeById(id);
+                } else {
+                    System.out.println("id не может быть отрицательным");
+                }
+            } catch (NumberFormatException e){
+                System.out.println("Неверный формат аргументов");
+            }
         }
     }
 
     public void clear(String arguments){
         if (!arguments.isBlank()){
-            System.out.println("гавно дибил"); // illegal args exception
+            System.out.println("Неверные аргументы команды"); // illegal args exception
         } else {
             collectionService.clear();
         }
@@ -106,37 +116,24 @@ public class CommandHandler {
 
     public void save(String arguments){
         if (!arguments.isBlank()){
-            System.out.println("гавно дибил"); // illegal args exception
+            System.out.println("Неверные аргументы команды"); // illegal args exception
         } else {
-            Scanner scanner = new Scanner(System.in);
-            System.out.println(
-                    "Введите имя файла в который хотите сохранить коллекцию"+ "\n" +
-                    "(по умолчанию файлу будет дано имя collection.csv)"
-            );
-            String filename = scanner.nextLine();
-            filename = filename.replaceAll(" ", "").replaceAll("\\.", "");
-            if (filename.isBlank()){
-                filename = "collection.csv";
-            } else {
-                filename += ".csv";
-            }
-            Path newFilePath = Paths.get(filename);
-            csvProvider.save(newFilePath, collectionService.collection);
+            csvProvider.save(collectionService.collection);
         }
     }
 
     public void executeScript(String path){  //args required
         if (path.isBlank()){
-            System.out.println("гавно дибил"); // illegal args exception
+            System.out.println("Неверные аргументы команды"); // illegal args exception
         } else {
-
+            // TODO: 19.02.2024 реализовать executeScript
         }
-        // TODO: 19.02.2024 реализовать executeScript
+
     }
 
     public void exit(String arguments){
         if (!arguments.isBlank()){
-            System.out.println("гавно дибил"); // illegal args exception
+            System.out.println("Неверные аргументы команды"); // illegal args exception
         } else {
             System.exit(0);
         }
@@ -144,16 +141,24 @@ public class CommandHandler {
 
     public void removeGreater(String arguments){ //args required
         if (arguments.isBlank()){
-            System.out.println("гавно дибил"); // illegal args exception
+            System.out.println("Неверные аргументы команды"); // illegal args exception
         } else {
-            long startId = Long.parseLong(arguments) + 1;
-            collectionService.removeGreater(startId);
+            try {
+                long startId = Long.parseLong(arguments) + 1;
+                if (startId > 0) {
+                    collectionService.removeGreater(startId);
+                } else {
+                    System.out.println("id не может быть отрицательным");
+                }
+            } catch (NumberFormatException e){
+                System.out.println("Неверный формат аргументов");
+            }
         }
     }
 
     public void reorder(String arguments){
         if (!arguments.isBlank()){
-            System.out.println("гавно дибил"); // illegal args exception
+            System.out.println("Неверные аргументы команды"); // illegal args exception
         } else {
             collectionService.reorder();
         }
@@ -161,7 +166,7 @@ public class CommandHandler {
 
     public void history(String arguments){
         if (!arguments.isBlank()){
-            System.out.println("гавно дибил"); // illegal args exception
+            System.out.println("Неверные аргументы команды"); // illegal args exception
         } else {
             System.out.println("Последние 7 команд, введённые пользователем: ");
             for (String command : commandList) {
@@ -172,7 +177,7 @@ public class CommandHandler {
 
     public void removeAllByType(String arguments){ //args required
         if (arguments.isBlank()){
-            System.out.println("гавно дибил"); // illegal args exception
+            System.out.println("Неверные аргументы команды"); // illegal args exception
         } else {
             try {
                 VehicleType type = VehicleType.valueOf(arguments.toUpperCase());
@@ -185,16 +190,24 @@ public class CommandHandler {
 
     public void countGreaterThanEnginePower(String arguments){ //args required
         if (arguments.isBlank()){
-            System.out.println("гавно дибил"); // illegal args exception
+            System.out.println("Неверные аргументы команды"); // illegal args exception
         } else {
-            double enginePower = Double.parseDouble(arguments);
-            collectionService.countGreaterThanEnginePower(enginePower);
+            try {
+                double enginePower = Double.parseDouble(arguments);
+                if (enginePower > 0) {
+                    collectionService.countGreaterThanEnginePower(enginePower);
+                } else {
+                    System.out.println("Мощность двигателя не может быть отрицательной");
+                }
+            } catch (NumberFormatException e){
+                System.out.println("Неверный формат аргументов");
+            }
         }
     }
 
     public void filterStartsWithName(String arguments){ //args required
         if (arguments.isBlank()){
-            System.out.println("гавно дибил"); // illegal args exception
+            System.out.println("Неверные аргументы команды"); // illegal args exception
         } else {
             collectionService.filterStartsWithName(arguments);
         }
